backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0

# Database
psycopg2-binary>=2.9,<3.0

# Environment & Configuration
django-environ>=0.11,<1.0

# API & Documentation
django-filter>=24.0,<25.0
drf-spectacular>=0.27,<1.0

# File Upload & Storage
Pillow>=10.0,<11.0
django-storages[s3]>=1.14,<2.0
boto3>=1.34,<2.0

# Async Tasks
celery>=5.3,<6.0
redis>=5.0,<6.0

# Payments
stripe>=8.0,<9.0

# Security & Utils
python-decouple>=3.8,<4.0
django-extensions>=3.2,<4.0

# Email
django-anymail[sendgrid]>=10.3,<11.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0

# Production
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0
